+++ zio-http-example/src/main/scala/example/HelloWorldWithMiddlewares.scala
@@ -20,7 +20,7 @@ object HelloWorldWithMiddlewares extends ZIOAppDefault {
   val serverTime: RequestHandlerMiddleware[Nothing, Any, Nothing, Any] = HttpAppMiddleware.patchZIO(_ =>
     for {
       currentMilliseconds <- Clock.currentTime(TimeUnit.MILLISECONDS)
-      withHeader = Patch.addHeader("X-Time", currentMilliseconds.toString)
+      withHeader = Response.Patch.addHeader("X-Time", currentMilliseconds.toString)
     } yield withHeader,
   )
   val middlewares =
