+++ zio-http/src/main/scala/zio/http/Response.scala
@@ -16,6 +16,8 @@
 
 package zio.http
 
+import scala.annotation.tailrec
+
 import zio._
 import zio.stacktracer.TracingImplicits.disableAutoTrace
 
@@ -90,8 +92,7 @@ sealed trait Response extends HeaderOps[Response] { self =>
     case _                    => false
   }
 
-  final def patch(p: Response.Patch): Response =
-    copy(headers = self.headers ++ p.addHeaders, status = p.setStatus.getOrElse(self.status))
+  final def patch(p: Response.Patch): Response = p.apply(self)
 
   private[zio] def serverTime: Boolean = false
 
@@ -259,9 +260,50 @@ object Response {
     }
   }
 
-  final case class Patch(addHeaders: Headers, setStatus: Option[Status]) { self =>
-    def ++(that: Patch): Patch =
-      Patch(self.addHeaders ++ that.addHeaders, self.setStatus.orElse(that.setStatus))
+  /**
+   * Models the set of operations that one would want to apply on a Response.
+   */
+  sealed trait Patch { self =>
+    def ++(that: Patch): Patch         = Patch.Combine(self, that)
+    def apply(res: Response): Response = {
+
+      @tailrec
+      def loop(res: Response, patch: Patch): Response =
+        patch match {
+          case Patch.Empty                  => res
+          case Patch.AddHeaders(headers)    => res.addHeaders(headers)
+          case Patch.RemoveHeaders(headers) => res.removeHeaders(headers)
+          case Patch.SetStatus(status)      => res.setStatus(status)
+          case Patch.Combine(self, other)   => loop(self(res), other)
+          case Patch.UpdateHeaders(f)       => res.updateHeaders(f)
+        }
+
+      loop(res, self)
+    }
+  }
+
+  object Patch {
+    import Header.HeaderType
+
+    case object Empty                                     extends Patch
+    final case class AddHeaders(headers: Headers)         extends Patch
+    final case class RemoveHeaders(headers: Set[String])  extends Patch
+    final case class SetStatus(status: Status)            extends Patch
+    final case class Combine(left: Patch, right: Patch)   extends Patch
+    final case class UpdateHeaders(f: Headers => Headers) extends Patch
+
+    def empty: Patch = Empty
+
+    def addHeader(headerType: HeaderType)(value: headerType.HeaderValue): Patch =
+      addHeader(headerType.name, headerType.render(value))
+
+    def addHeader(header: Header): Patch                          = addHeaders(Headers(header))
+    def addHeaders(headers: Headers): Patch                       = AddHeaders(headers)
+    def addHeader(name: CharSequence, value: CharSequence): Patch = addHeaders(Headers(name, value))
+
+    def removeHeaders(headerTypes: Set[HeaderType]): Patch = RemoveHeaders(headerTypes.map(_.name))
+    def setStatus(status: Status): Patch                   = SetStatus(status)
+    def updateHeaders(f: Headers => Headers): Patch        = UpdateHeaders(f)
   }
 
   def apply(
